Add files via upload
Customer Segmentation Based on Credit Risk

About Dataset
Context
The original dataset contains 1000 entries with 20 categorial/symbolic attributes prepared by Prof.
Hofmann. In this dataset, each entry represents a person who takes a credit by a bank. Each person
is classified as good or bad credit risks according to the set of attributes. The link to the original
dataset can be found below.
Content
It is almost impossible to understand the original dataset due to its complicated system of categories
and symbols. Thus, I wrote a small Python script to convert it into a readable CSV file. Several
columns are simply ignored, because in my opinion either they are not important or their descriptions
are obscure. The selected attributes are:

1. Age (numeric)
2. Sex (text: male, female)
3. Job (numeric: 0 - unskilled and non-resident, 1 - unskilled and resident, 2 - skilled, 3 - highly
skilled)
4. Housing (text: own, rent, or free)
5. Saving accounts (text - little, moderate, quite rich, rich)
6. Checking account (numeric, in DM - Deutsch Mark)
7. Credit amount (numeric, in DM)
8. Duration (numeric, in month)
9. Purpose (text: car, furniture/equipment, radio/TV, domestic appliances, repairs, education,
business, vacation/others)

Introduction
&quot;Unlocking Financial Insights:
A Comprehensive Data Analysis of Credit Information for Evaluating Applicant Creditworthiness&quot;
In today&#39;s rapidly evolving financial landscape, accurate and robust credit assessment is paramount
for informed lending decisions. This data-driven analysis delves into the intricate web of credit data
to uncover patterns, trends, and key indicators that contribute to determining the creditworthiness of
applicants.

● Not approving a loan to an applicant having a good credit risk is a loss of business to the
bank.
● Approving a loan to an applicant having a bad credit risk is also a loss of business to the
bank.
Hence, banks have to be very careful while accepting or rejecting a loan application.
Objective
The main objective of this assignment is to minimize the risk and maximize the profit of the bank.
Bank has to make a decision based on the applicant’s profile to minimize the loss from the bank&#39;s
perspective.
The task is to create a Machine Learning model that can predict the credit risk assessment for a
given application. Due to the binary nature of the prediction (good or bad), this is a Binary
Classification problem consisting of 3 numerical variables and 6 categorical variables, excluding the
target column. The need for combining these fundamentally different variables and exploring hidden
relationships between them, make all following steps equally crucial and relevant.
1-) Exploratory Data Analysis
2-) Data Preprocessing
3-) Machine Learning